+++ toteatds/src/commonMain/kotlin/com/toteat/toteatds/theme/Color.kt
@@ -18,6 +18,12 @@ val NeutralGray200 = Color(0xFFDDDDDD)
 val NeutralGray100 = Color(0xFFF2F2F2)
 val NeutralGray = Color(0xFFFFFFFF)
 
+// Counter Colors
+val CounterContainerColor = Color(0xFFF7F4F0)
+val CounterContainerDisabledColor = Color(0xFFF8F8F8)
+val CounterButtonColor = Color(0xFFE3DBCC)
+val CounterButtonDisabledColor = Color(0xFFF8F8F8)
+
 // Feedback / tags
 val GreenNormal = Color(0xFF46D281)
 val GreenLight = Color(0xFFCDF3DB)